- Added: Ability to rollback to a previous version - Added: Updater UI which can install or update to any SMA version (rescue tool), when no arguments are passed to Update.exe - Added: Fody PopertyChanged Package - Added: Fody Costura Package (replaces ILRepack) - Added: ConfigureAwait (WIP) - Added: CalculateUpdatePath and FetchAllReleases to the UpdateManager API - Added: Update.exe is now reweaved during --releasify with the project's PackageName, Title, and optional update URL for the GUI - Added: Test cases for the new features - Fixed: Several medium & minor issues - Fixed: Issue finding tools depending on environment - Updated: Stub executable is now a .Net45 project - Updated: Stub executable now selects the version based on the local RELEASES file first - Updated: Old Stub executable project is renamed to legacy (swap it back to use instead of the new StubExecutable if desired) - Updated: Improved Installer ability to optimize the download size by choosing delta updates - Updated: Refactored the project to improve the organization - Updated: Improved the tests code quality and style (WIP) - Updated: Test cases to function with the improvements - Updated: Packages version

+++ test/ReleasePackageTests.cs
@@ -10,6 +10,7 @@ using System.Linq;
 using System.Reflection;
 using System.Xml.Linq;
 using Splat;
+using Squirrel.Extensions;
 using Xunit;
 
 namespace Squirrel.Tests.Core
@@ -22,7 +23,7 @@ namespace Squirrel.Tests.Core
             var sv = new SemanticVersion("1.2.3.4");
             var dontcare = default(SemanticVersion);
 
-            Assert.False(SemanticVersion.TryParseStrict(sv.ToString(), out dontcare));
+            Assert.True(SemanticVersion.TryParseStrict(sv.ToString(), out dontcare));
         }
 
         [Fact]
@@ -216,7 +217,9 @@ namespace Squirrel.Tests.Core
                 Assert.Equal(5, assemblyName.Version.Minor);
             } finally {
                 File.Delete(outputPackage);
-                File.Delete(outputFile);
+
+                if (outputFile!= null)
+                  File.Delete(outputFile);
             }
         }
